+++ WeatherAppAssessment/Feature/Home/view/HomeView.swift
@@ -14,11 +14,6 @@ import MapKit
 // add a serach icon to allow searching cities
 
 struct HomeView: View {
-    struct CoordinateSelection: Identifiable {
-        let id = UUID()
-        let coordinates: CLLocationCoordinate2D
-    }
-    
     @Environment(LocationViewModel.self) private var locationVM
     
     @State private var weatherVM = WeatherViewModel()
@@ -63,17 +58,15 @@ struct HomeView: View {
                             Image(systemName: "map")
                                 .imageScale(.large)
                         }
-        //                Text("Search")
-        //                Label("Search", systemImage: "magnifyingglass")
                     }
                     .foregroundStyle(.white)
                     .font(.title3.weight(.medium))
                     .padding()
                     .background(Material.ultraThin.opacity(0.4))
                 }
-                .sheet(isPresented: $presentExtraInfo, content: {
-                    extraInfoSheet(weather)
-                })
+                .sheet(isPresented: $presentExtraInfo) {
+                    ExtraInformationSheet(weather: weather)
+                }
             case .error(let description):
                 ContentUnavailableView("Something Went Wrong", systemImage: "exclamationmark.triangle", description: Text("\(description). Please try again"))
             }
@@ -81,9 +74,9 @@ struct HomeView: View {
         .sheet(isPresented: $presentFavorites) {
             FavoritesList()
         }
-        .sheet(item: $mapSelection, content: { selection in
+        .sheet(item: $mapSelection) { selection in
             MapView(location: selection.coordinates)
-        })
+        }
         .task {
             self.location = locationVM.getUserLocation()
         }
@@ -117,55 +110,6 @@ struct HomeView: View {
         case "clear": .clear
         case "partlysunny": .partlysunny
         default: .clear
-        }
-    }
-    
-    @ViewBuilder
-    func extraInfoSheet(_ weather: CurrentWeather) -> some View {
-        VStack(alignment: .leading) {
-            VStack(alignment: .leading) {
-                Text(weather.name)
-                    .font(.title)
-                Text(weather.country)
-                    .font(.title3)
-            }
-            
-            VStack(alignment: .leading) {
-                Image(systemName: "mappin")
-                LabeledContent("Longitude", value: weather.coordinates.longitude.formatted())
-                LabeledContent("Latitude", value: weather.coordinates.latitude.formatted())
-            }
-            .padding(.vertical)
-            
-            VStack(alignment: .leading) {
-                Image(systemName: "sun.horizon")
-                LabeledContent("Sunrise", value: weather.sunrise.formatted(.dateTime.hour().minute()))
-                LabeledContent("Sunset", value: weather.sunset.formatted(.dateTime.hour().minute()))
-            }
-            .padding(.vertical)
-
-            Text("Last updated: \(Date(), format: .reference(to: weather.date))")
-        }
-//        .foregroundStyle(.white)
-        .fontWeight(.medium)
-//        .presentationBackground(backgroundColor(weather.description))
-        .presentationDetents([.medium])
-        .presentationCornerRadius(32)
-        .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-        .padding(20)
-        .overlay(alignment: .topTrailing) {
-            Button {
-                presentExtraInfo = false
-            } label: {
-                Image(systemName: "plus")
-                    .imageScale(.large)
-                    .rotationEffect(.degrees(45))
-            }
-            .fontWeight(.medium)
-            .foregroundStyle(.primary)
-            .buttonStyle(.bordered)
-            .buttonBorderShape(.circle)
-            .padding()
         }
     }
